chore(src): reorganize drawer element
