+++ src/helix-ui/elements/hx-drawer/_shadow.less
@@ -1,4 +1,4 @@
-@import "./HXElement";
+@import "../HXElement";
 @import (reference) "components/buttons";
 
 #hxClose {
