+++ src/helix-ui/elements/hx-drawer/index.js
@@ -1,6 +1,7 @@
-import { HXElement } from './HXElement';
-import shadowMarkup from './HXDrawerElement.html';
-import shadowStyles from './HXDrawerElement.less';
+import { HXElement } from '../HXElement.js';
+
+import shadowMarkup from './_shadow.html';
+import shadowStyles from './_shadow.less';
 
 /**
  * Cancelable event that fires when the element's
