+++ src/helix-ui/elements/index.js
@@ -7,7 +7,7 @@ export { HXCheckboxElement } from './hx-checkbox/index.js';
 export { HXCheckboxSetElement } from './hx-checkbox-set/index.js';
 export { HXDisclosureElement } from './hx-disclosure/index.js';
 export { HXDivElement } from './hx-div/index.js';
-export { HXDrawerElement } from './HXDrawerElement';
+export { HXDrawerElement } from './hx-drawer/index.js';
 export { HXDropFenceElement } from './hx-drop-fence/index.js';
 export { HXDropZoneElement } from './hx-drop-zone/index.js';
 export { HXElement } from './HXElement';
